requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.8
diploma-frontend==0.6
Django==5.0.6
django-filter==24.2
django-request-logging==0.7.5
djangorestframework==3.15.1
djangorestframework-filters==0.11.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
pillow==10.3.0
pycparser==2.22
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
